fix regression with converting id strings to object id
