+++ src/model.js
@@ -277,6 +277,7 @@ const customMethods = [
 	'exclude',
 	'search',
 	'remove',
+	'where',
 	'count',
 	'find',
 	'sort'
